+++ subs-aeagent/src/main/java/uk/ac/ebi/subs/arrayexpress/model/ArrayExpressStudy.java
@@ -2,7 +2,7 @@ package uk.ac.ebi.subs.arrayexpress.model;
 
 
 import org.springframework.data.annotation.Id;
-import uk.ac.ebi.subs.data.component.Protocol;
+import uk.ac.ebi.subs.data.submittable.Protocol;
 import uk.ac.ebi.subs.data.submittable.Study;
 
 import java.util.ArrayList;
