+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/component/ProtocolRef.java
@@ -1,5 +1,7 @@
 package uk.ac.ebi.subs.data.component;
 
+import uk.ac.ebi.subs.data.submittable.Protocol;
+
 public class ProtocolRef extends AbstractSubsRef<Protocol> {
     Protocol referencedObject;
